Defend against argument inversion on the commandline.
This is about the only way in which you can fail to use the API, because
the public and private keys are indistinguishable 32-byte blocks. So,
for storage, we'll tack on a "type" trailer, and check it back on use.

This won't defend against people willfully changing the key types under
the hood, but all I'm aiming for here is preventing accidental misuse.